+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_client.py
@@ -24,9 +24,10 @@ from typing import MutableMapping
 
 import grpc
 
-from opentelemetry import trace
+from opentelemetry import context, trace
 from opentelemetry.instrumentation.grpc import grpcext
 from opentelemetry.instrumentation.grpc._utilities import RpcInfo
+from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.propagate import inject
 from opentelemetry.propagators.textmap import Setter
 from opentelemetry.semconv.trace import SpanAttributes
@@ -101,6 +102,9 @@ class OpenTelemetryClientInterceptor(
         return result
 
     def _intercept(self, request, metadata, client_info, invoker):
+        if context.get_value(_SUPPRESS_INSTRUMENTATION_KEY):
+            return invoker(request, metadata)
+
         if not metadata:
             mutable_metadata = OrderedDict()
         else:
@@ -184,6 +188,9 @@ class OpenTelemetryClientInterceptor(
     def intercept_stream(
         self, request_or_iterator, metadata, client_info, invoker
     ):
+        if context.get_value(_SUPPRESS_INSTRUMENTATION_KEY):
+            return invoker(request_or_iterator, metadata)
+
         if client_info.is_server_stream:
             return self._intercept_server_stream(
                 request_or_iterator, metadata, client_info, invoker