+++ instrumentation/opentelemetry-instrumentation-grpc/tests/test_client_interceptor.py
@@ -18,7 +18,7 @@ from tests.protobuf import (  # pylint: disable=no-name-in-module
 )
 
 import opentelemetry.instrumentation.grpc
-from opentelemetry import trace
+from opentelemetry import context, trace
 from opentelemetry.instrumentation.grpc import GrpcInstrumentorClient
 from opentelemetry.instrumentation.grpc._client import (
     OpenTelemetryClientInterceptor,
@@ -26,6 +26,7 @@ from opentelemetry.instrumentation.grpc._client import (
 from opentelemetry.instrumentation.grpc.grpcext._interceptor import (
     _UnaryClientInfo,
 )
+from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
 from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.mock_textmap import MockTextMapPropagator
@@ -301,3 +302,47 @@ class TestClientProto(TestBase):
 
         finally:
             set_global_textmap(previous_propagator)
+
+    def test_unary_unary_with_suppress_key(self):
+        token = context.attach(
+            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
+        )
+        try:
+            simple_method(self._stub)
+            spans = self.memory_exporter.get_finished_spans()
+        finally:
+            context.detach(token)
+        self.assertEqual(len(spans), 0)
+
+    def test_unary_stream_with_suppress_key(self):
+        token = context.attach(
+            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
+        )
+        try:
+            server_streaming_method(self._stub)
+            spans = self.memory_exporter.get_finished_spans()
+        finally:
+            context.detach(token)
+        self.assertEqual(len(spans), 0)
+
+    def test_stream_unary_with_suppress_key(self):
+        token = context.attach(
+            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
+        )
+        try:
+            client_streaming_method(self._stub)
+            spans = self.memory_exporter.get_finished_spans()
+        finally:
+            context.detach(token)
+        self.assertEqual(len(spans), 0)
+
+    def test_stream_stream_with_suppress_key(self):
+        token = context.attach(
+            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
+        )
+        try:
+            bidirectional_streaming_method(self._stub)
+            spans = self.memory_exporter.get_finished_spans()
+        finally:
+            context.detach(token)
+        self.assertEqual(len(spans), 0)